expanded on crucible stuff, added more docs and WIP cosmetic system

+++ src/main/java/io/github/vampirestudios/obsidian/api/crucible/CrucibleSkill.java
@@ -8,11 +8,15 @@ import java.util.List;
 public class CrucibleSkill {
     public Identifier id;
 
-    public int Cooldown;
+    // YAML uses 80.0 etc -> use double
+    public double Cooldown;
+
     public List<String> Skills;
     public List<String> Conditions;
     public List<String> TargetConditions;
     public List<String> TriggerConditions;
+
+    // Compiled steps (one per line) OR could be left empty if you only keep the wrapper
     public List<Skill> internalSkills;
 
     public CrucibleSkill() {
